test: added test for the signup ui components
