--- test/unit/register/hybee-signup-test-2/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "functionality-test",
-  "version": "1.0.0",
-  "description": "",
-  "main": "util.js",
-  "scripts": {
-    "test": "mocha"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "chai": "^4.2.0",
-    "mocha": "^7.2.0"
-  }
-}
