--- test/unit/register/hybee-signup-test-2/util.js
@@ -1,77 +1,0 @@
-const validateInput = (text, notEmpty, isNumber) => {
-  // Validate user input with two pre-defined rules
-  if (!text) {
-    return false;
-  }
-  if (notEmpty && text.trim().length === 0) {
-    return false;
-  }
-  if (isNumber && +text === NaN) {
-    return false;
-  }
-  return true;
-};
-// const validateNumber = (number) => {
-//   // Validate user input with two pre-defined rules
-//   if (!isNaN(number)) {
-//     return false;
-//   }
-//   return true;
-// };
-
-const validateEmail = (email, notEmpty) => {
-  if (!email) {
-    return false;
-  }
-  if (notEmpty && email.trim().length === 0) {
-    return false;
-  }
-  return true;
-};
-
-const getField = (name) => {
-  // Returns output text
-  return `${name}`;
-};
-
-const validatePassword = () => {};
-
-// const checkField =
-
-const functions = {
-  checkName: (name) => {
-    if (!validateInput(name, true, false)) {
-      return false;
-    }
-
-    return `${getField(name)}`;
-  },
-
-  checkEmail: (email) => {
-    if (!validateEmail(email, true)) {
-      return false;
-    }
-
-    return `${getField(email)}`;
-  },
-
-  checkPassword: (password) => {
-    if (!validateInput(password, true, false)) {
-      return false;
-    }
-
-    return `${getField(password)}`;
-  },
-
-  confirmPassword: (passwordConfirm) => {
-    if (!validateInput(passwordConfirm, true, false)) {
-      return false;
-    }
-
-    return `${getField(passwordConfirm)}`;
-  },
-};
-
-// const checkNames =
-
-module.exports = functions;
